Real coupling test

+++ utils/engines_handler.py
@@ -19,6 +19,11 @@ class EnginesHandler:
     coupler only one instance of the EngineHandler class is needed.
     """
     def __init__(self, engine):
+
+        # Check input parameters
+        if not isinstance(engine, str):
+            raise TypeError(f'Input parameter engine must be of type string is {type(engine)}.')
+
         self.engine_name = engine
         self.engine = None
         self.iterations = []
@@ -27,13 +32,13 @@ class EnginesHandler:
 
         self.log = log.getLogger(f'{self.__class__.__name__}:{self.engine_name}')
 
-    def init_engine(self, parameter: dict = {}):
+    def init_engine(self, parameter: dict = None):
         """
         Initialize engine according to self.engine_name. Parameters needed to initialize any engine differ
         on the engine and need to be of type dictionary.
 
         Args:
-            parameter (dict): Dictionary including parameters
+            parameter (dict, optional): Dictionary including parameters
 
         Returns:
             specific engine: for example AbaqusEngine()
@@ -43,47 +48,74 @@ class EnginesHandler:
             self.log.error(f'Engine already initialized: {self.engine}')
             return False
 
+        if not isinstance(parameter, dict) and parameter is not None:
+            raise TypeError(f'Input parameter parameter must be of type Dictionary is {type(parameter)}.')
+
+        if parameter is None:
+            parameter = {}
+
+        # Initialization of an engine depends on the engine type.
         if self.engine_name == 'abaqus':
             if 'input_file' in parameter:
                 engine = AbaqusEngine(parameter['input_file'])
             else:
-                self.log.error(f'Missing parameter "input_file" in parameters: {parameter}')
-                raise TypeError
+                raise TypeError(f'Missing parameter "input_file" in parameters: {parameter}')
 
         elif self.engine_name == 'pace3d':
             engine = Pace3dEngine()
 
         else:
-            self.log.error(f'Engine {self.engine_name} is unknown.')
-            raise TypeError
+            raise TypeError(f'Engine {self.engine_name} is unknown.')
 
         self.log.info(f'Engine {self.engine_name} successfully initialized')
         self.engine = engine
         return engine
 
-    def add_iteration_step(self, iteration_name, previous_copy=False):
+    def add_iteration_step(self, iteration_name, previous_copy=False, delete_previous_grid=True):
         """ Add an iteration step to this instance of the engine handler.
 
         Args:
-            iteration_name: name of the iteration to be added
-            previous_copy (boolean): set true to make a (deep)copy of the previous step, keeping grid information
-                                     and values
+            iteration_name (str): name of the iteration to be added
+            previous_copy (bool, optional): set true to make a (deep)copy of the previous step, keeping grid
+                                                information and values
+            delete_previous_grid (bool, optional): set true to delete grid from previous iteration step to save
+                                                        memory
 
         Returns:
             IterationStep on success
         """
 
+        # Check input parameters
+        if not isinstance(iteration_name, str):
+            raise TypeError(f'Input parameter iteration_name must be of type string is {type(iteration_name)}.')
+        if not isinstance(previous_copy, bool):
+            raise TypeError(f'Input parameter previous_copy must be of type boolean is {type(previous_copy)}.')
+        if not isinstance(delete_previous_grid, bool):
+            raise TypeError(f'Input parameter delete_previous_grid must be of type boolean is '
+                            f'{type(delete_previous_grid)}.')
+
         step_no = len(self.iterations)
         step_name_exists = False
 
+        # Check if the step name given by a input parameter is used for a previous step. The name must be unique. There
+        # fore a error is raised, when the name already exists.
         for iterations in self.iterations:
             if iterations.name == iteration_name:
                 step_name_exists = True
 
         if step_name_exists:
-            self.log.error(f'Step with the same name ({iteration_name}) already exists.')
-            raise KeyError
+            raise NameError(f'Step with the same name ({iteration_name}) already exists.')
+
         else:
+            # Delete grid of previous simulation. This step is only necessary to save systems memory. The grid of the
+            # actual step will be kept and only the grid of the previous step will be deleted. Steps before the previous
+            # step are not affected.
+            if delete_previous_grid:
+                if len(self.iterations) > 1:
+                    self.iterations[len(self.iterations) - 2].grid = None
+
+            # If input parameter previous_copy is set a deepcopy of the previous step will be created and only the
+            # name and step_no will be changed.
             if previous_copy:
                 previous_iteration = self.iterations[len(self.iterations) - 1]
                 current_iteration = copy.deepcopy(previous_iteration)
@@ -96,7 +128,7 @@ class EnginesHandler:
 
                 self.log.debug(f'Added copy of previous iteration step "{previous_iteration_name}" with new '
                                f'name "{iteration_name}"')
-                return current_iteration  # self.iterations[step_no]
+                return current_iteration
             else:
                 self.iterations.append(IterationStep(iteration_name, step_no))
                 self.log.debug(f'Added new iteration step: {iteration_name}')
@@ -108,11 +140,16 @@ class EnginesHandler:
         Returns:
             IterationStep
         """
-        return self.iterations[len(self.iterations) - 1]
+        if len(self.iterations) > 0:
+            return self.iterations[len(self.iterations) - 1]
+        else:
+            self.log.warning(f'There is not iteration step to get. First an iteration has to be initialized by '
+                             f'add_iteration_step()')
+            return False
 
     def path_cleanup(self, path_name, recreate_missing=True):
         """
-        Cleaning up the given path means deleting all files and subfolder. Folders saved in self.path will be checked
+        Cleaning up the given path means deleting all files and sub folder. Folders saved in self.path will be checked
         and if needed and parameter recreate_missing is set be recreated. No longer existing files in self.files will
         be deleted from dictionary.
 
@@ -123,30 +160,31 @@ class EnginesHandler:
         Returns:
             boolean: true on success
         """
+
+        # Check input parameter
+        if not isinstance(recreate_missing, bool):
+            raise TypeError(f'Input parameter recreate_missing must be of type boolean, is {type(recreate_missing)}.')
+
         path = self.get_path(path_name)
 
         # Check if path is file or folder
         if path.is_file():
-            self.log.error(f'Given path is a file. Expected path.')
-            raise TypeError
+            raise TypeError(f'Given path is a file. Expected path.')
 
         # Remove recursive and optionally recreate folder (see recreate_missing option)
         if path.is_dir():
-            shutil.rmtree(path) # Remove all files and sub folders recursively
-            # time.sleep(0.5)
+            shutil.rmtree(path)  # Remove all files and sub folders recursively
             path.mkdir()  # Create dir
             self.log.info(f'Cleaned up path {path}')
 
-        # FIXME Recreate only betroffene folder
         if recreate_missing:
-            print('TEST:', path.parent)
-            print('TEST:', path.parents)
             # Check if all paths in self.path are exist, otherwise create.
             for name, path in self.paths.items():
-                if not path.is_dir():
-                    time.sleep(0.5)
-                    path.mkdir()
-                    self.log.info(f'Recreated path for {name} at {path}')
+                if name == path_name:
+                    if not path.is_dir():
+                        time.sleep(0.5)
+                        path.mkdir()
+                        self.log.info(f'Recreated path for {name} at {path}.')
 
         # Check if all files still exist, otherwise delete from self.files
         remove_names = []
@@ -174,38 +212,40 @@ class EnginesHandler:
         Returns:
             boolean: true on success
         """
-        try:
-            path = Path(path)
 
-            if path.is_file():
-                self.log.error(f'Given path {path} is a file. Use .set_file() instead.')
-                raise TypeError
+        # Check input parameters
+        if not isinstance(path_name, str):
+            raise TypeError(f'Input parameter path_name must be of type string, is {type(path_name)}.')
+        if not isinstance(path, str) and not isinstance(path, Path):
+            raise TypeError(f'Input parameter path must be of type string or Path, is {type(create_missing)}.')
+        if not isinstance(create_missing, bool):
+            raise TypeError(f'Input parameter create_missing must be of type boolean, is {type(create_missing)}.')
 
-            if not path.is_dir():
-                if create_missing:
-                    path.mkdir()
+        path = Path(path)
 
-                    if path.is_dir():
-                        self.log.debug(f'Path {path_name} generated:{path}')
-                        self.paths[path_name] = path
-                        self.log.debug(f'Checked and added path {path_name} : {path}')
-                        return True
+        if path.is_file():
+            raise NotADirectoryError(f'Given path {path} is a file. Use .set_file() instead.')
 
-                    else:
-                        self.log.error(f'Not able to create path {path_name} : {path}')
+        if not path.is_dir():
+            if create_missing:
+                path.mkdir()
+
+                if path.is_dir():
+                    self.log.debug(f'Path {path_name} generated:{path}')
+                    self.paths[path_name] = path
+                    self.log.debug(f'Checked and added path {path_name} : {path}')
+                    return True
 
                 else:
-                    self.log.warning(f'Path {path} does not exist.')
-                    raise FileNotFoundError
+                    raise OSError(f'Not able to create path {path_name} : {path}')
 
             else:
-                self.paths[path_name] = path
-                self.log.debug(f'Checked and added path {path_name} : {path}')
-                return True
+                raise NotADirectoryError(f'Path {path} does not exist.')
 
-        except FileNotFoundError as err:
-            self.log.error(f'Error occured while setting path {path} as {path_name}. {err}')
-            raise FileNotFoundError
+        else:
+            self.paths[path_name] = path
+            self.log.debug(f'Checked and added path {path_name} : {path}')
+            return True
 
     def get_path(self, path_name):
         """
@@ -217,12 +257,22 @@ class EnginesHandler:
         Returns:
             Path: Path of the path
         """
-        try:
-            return self.paths[path_name]
 
-        except KeyError as err:
-            self.log.error(f'Error occurred while reading path {path_name}. {err}')
-            raise KeyError
+        # Check input parameters
+        if not isinstance(path_name, str):
+            raise TypeError(f'Input parameter path_name must be of type string, is {type(path_name)}.')
+        else:
+            # Check if path_name is part of self.paths
+            path_name_exists = False
+
+            for key in self.paths.keys():
+                if key == path_name:
+                    path_name_exists = True
+
+            if not path_name_exists:
+                raise KeyError(f'Path name does not exist. ({path_name})')
+
+        return self.paths[path_name]
 
     def set_file(self, file_name, file):
         """
@@ -230,28 +280,28 @@ class EnginesHandler:
 
         Args:
             file_name (str):  Name of the file for later use
-            file (str, path): File to add as Path or String
+            file (str/path): File to add as Path or String
 
         Returns:
             boolean: true on success
         """
-        try:
-            file = Path(file)
-            if file.is_dir():
-                self.log.error(f'Given file is a path. Use .set_path() instead.')
-                raise IsADirectoryError
 
-            if file.is_file():
-                self.files[file_name] = file
-                self.log.debug(f'Checked and added file {file_name} : {file}')
-                return True
-            else:
-                self.log.error(f'File {file} not found.')
-                raise FileNotFoundError
+        # Check input parameters
+        if not isinstance(file_name, str):
+            raise TypeError(f'Input parameter file_name must be of type string, is {type(file_name)}.')
+        if not isinstance(file, str) and not isinstance(file, Path):
+            raise TypeError(f'Input parameter file must be of type string or Path, is {type(file_name)}.')
 
-        except FileNotFoundError as err:
-            self.log.error(f'Error occurred while setting file {file} as {file_name}. {err}')
-            raise FileNotFoundError
+        file = Path(file)
+        if file.is_dir():
+            raise IsADirectoryError(f'Given file is a directory. Use .set_path() instead.')
+
+        if file.is_file():
+            self.files[file_name] = file
+            self.log.debug(f'Checked and added file {file_name} : {file}')
+            return True
+        else:
+            raise FileNotFoundError(f'File not found. ({file})')
 
     def get_file(self, file_name):
         """
@@ -263,9 +313,19 @@ class EnginesHandler:
         Returns:
             Path: Path to file
         """
-        try:
-            return self.files[file_name]
 
-        except KeyError as err:
-            self.log.error(f'Error occured while reading file {file_name}. {err}')
-            raise KeyError
+        # Check input parameters
+        if not isinstance(file_name, str):
+            raise TypeError(f'Input parameter file_name must be of type string, is {type(file_name)}.')
+        else:
+            # Check if path_name is part of self.paths
+            file_name_exists = False
+
+            for key in self.files.keys():
+                if key == file_name:
+                    file_name_exists = True
+
+            if not file_name_exists:
+                raise KeyError(f'Path name does not exist. ({file_name})')
+
+        return self.files[file_name]
